ASCRAPPY-6 Removed Modal and added Page
