+++ src/app/home/home.component.html
@@ -15,18 +15,6 @@
     </p>
 
     <p class="text-secondary">
-        In addition to our news coverage, we offer a variety of web handy utilities to enhance your online experience.
-        Whether
-        you're looking for practical tools, entertainment, or productivity boosters, we have you covered. Discover a
-        collection
-        of useful utilities such as language translators, currency converters, weather updates, time zone calculators,
-        and more.
-        Stay organized with our note-taking and task management tools, and find inspiration with our creative resources
-        like
-        image galleries and quote libraries.
-    </p>
-
-    <p class="text-secondary">
         This website is designed with user-friendliness in mind, providing a seamless and intuitive browsing experience.
         With a
         clean and modern interface, you can easily navigate through different sections, access news articles, and
