+++ src/app/news/news.component.html
@@ -57,8 +57,8 @@
                                 <br>
                                 <div class="row">
                                     <div class="col">
-                                        <button class="btn btn-outline-dark mt-3" data-bs-toggle="modal"
-                                            data-bs-target="#myModal" (click)="readMore(n)">Read
+                                        <button class="btn btn-outline-dark mt-3"
+                                            [routerLink]="['/news', n.news ,selectedNewsService, n.url]">Read
                                             More</button>
                                     </div>
                                     <div class="col">
@@ -79,5 +79,4 @@
         <br>
     </div>
     <br>
-    <app-news-modal [newsTitle]="newsTitle" [newsContents]="newsContents"></app-news-modal>
 </div>
